requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Django==4.2.6
django-cors-headers==4.2.0
django-rq==2.8.1
djangorestframework==3.14.0
dnspython==2.4.2
idna==3.4
pymongo==4.5.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
rq==1.15.1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.6
